initial
some overview facts to the problems framework in relation covid19/transparency as a essential example of a law or rule-set decribing paths and abilities for more and clearer and also for everybody verivicable data and informations from public sector in the thirst and from sectors with highly influencing potencial in real social impacts ( like e.g. social media platforms).